pages/requirements.txt
pandas==2.2.3
protobuf==6.31.0
st-gsheets-connection==0.1.0
streamlit==1.45.1
google-genai
